First runable program - Moved all classes to the console app folder - Added missing namespace to TollCalculator class - Removed unncessary `using` directive - Added a README file with instructions to develop and run - Used C# 10/.NET 6 file-scoped namespaces for cleaner syntax
