+++ C#/C#.sln
@@ -1,0 +1,24 @@
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.5.2.0
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TollFeeCalculator", "TollFeeCalculator\TollFeeCalculator.csproj", "{49107A91-51F9-E53F-58E1-D2FF449669C7}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{49107A91-51F9-E53F-58E1-D2FF449669C7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{49107A91-51F9-E53F-58E1-D2FF449669C7}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{49107A91-51F9-E53F-58E1-D2FF449669C7}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{49107A91-51F9-E53F-58E1-D2FF449669C7}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {4D42EC99-EBC3-4095-8BE1-8D3E82B046B7}
+	EndGlobalSection
+EndGlobal
