+++ C#/README.md
@@ -1,0 +1,16 @@
+# Toll Fee Calculator C#
+
+This readme is intented to give instructions on how to setup and run the program.
+
+## Prerequisites
+
+To develop and run the program you will need the following
+
+1. .NET 8 SDK from [https://dotnet.microsoft.com/en-us/download/dotnet/8.0](here)
+2. IDE of your choice. If you are using VSCode please check the `.vscode/extensions.json` file for the required extensions
+
+## Dev & Run
+
+To run the app from the cli
+1. Make sure that you are in the directgory `C#/TollFreeCalculator`
+2. Run the app using `dotnet run`
