+++ C#/TollFeeCalculator/Car.cs
@@ -1,0 +1,10 @@
+﻿namespace TollFeeCalculator
+{
+    public class Car : Vehicle
+    {
+        public String GetVehicleType()
+        {
+            return "Car";
+        }
+    }
+}
