+++ C#/TollFeeCalculator/Motorbike.cs
@@ -1,0 +1,9 @@
+﻿namespace TollFeeCalculator;
+
+public class Motorbike : Vehicle
+{
+    public string GetVehicleType()
+    {
+        return "Motorbike";
+    }
+}
