+++ C#/TollFeeCalculator/Vehicle.cs
@@ -1,0 +1,6 @@
+﻿namespace TollFeeCalculator;
+
+public interface Vehicle
+{
+    String GetVehicleType();
+}
